refactor: Message 도메인과 정규화 로직 분리
- Raw → Message 변환 로직을 internal/messageNormalizer.ts로 이동
- 도메인 모델과 데이터 정규화 책임을 분리하여 역할을 명확히 함
- useChat / useVoiceChat에서 재사용 가능하도록 구조 개선
